+++ swap/src/network/swarm.rs
@@ -44,6 +44,7 @@ pub fn asb<LR>(
     wormhole_num_intro_points: u8,
     wormhole_swap_freshness_hours: u64,
     trust_provider: Arc<dyn super::wormhole::PeerTrust + Send + Sync>,
+    connection_limit_exemption_freshness_days: u64,
     metrics_registry: Option<&mut Registry>,
 ) -> Result<(
     Swarm<asb::Behaviour<LR>>,
@@ -105,6 +106,7 @@ where
         rendezvous_nodes,
         connection_limits,
         trust_provider,
+        connection_limit_exemption_freshness_days,
         // Passing None disables the wormhole behaviour entirely.
         if wormhole_enabled {
             wormhole_channels